refactor(supplier-profile): extract ProductFormStepProgress into its own file
